requirements.txt
opencc-python-reimplemented
tkinterdnd2